+++ lib/moqx/native.ex
@@ -18,8 +18,16 @@ defmodule MOQX.Native do
   def finish_track(_track), do: :erlang.nif_error(:nif_not_loaded)
 
   # Subscribe
-  def subscribe(_session, _broadcast_path, _track_name, _delivery_timeout_ms),
-    do: :erlang.nif_error(:nif_not_loaded)
+  def subscribe(
+        _session,
+        _subscription_ref,
+        _broadcast_path,
+        _track_name,
+        _delivery_timeout_ms,
+        _init_data,
+        _track_meta
+      ),
+      do: :erlang.nif_error(:nif_not_loaded)
 
   # Fetch
   def fetch(_session, _ref, _namespace, _track_name, _priority, _group_order, _start, _end),